+++ app/validators/date_within_boundaries_validator.rb
@@ -8,7 +8,7 @@ class DateWithinBoundariesValidator < ActiveModel::EachValidator
     unless value.between?(MIN.years.ago, MAX.years.from_now)
       record.errors.add(
         attribute,
-        I18n.t("activerecord.errors.models.#{record.class.name.downcase}.attributes.#{attribute}.between", min: MIN, max: MAX)
+        I18n.t("activerecord.errors.models.#{record.class.name.underscore.downcase}.attributes.#{attribute}.between", min: MIN, max: MAX)
       )
     end
   end