+++ CHANGELOG.md
@@ -139,6 +139,7 @@
 - Planned disbursements are exported in the IATI XML
 - The new planned disbursement form pre-fills the providing organisation details
 - Add cookie policy
+- Planned disbursement dates are validated within boundaries
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-6...HEAD
 [release-6]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-5...release-6